Refactor:Extract rendered classRegister... into own component;Extract AddPrevClassStudent checkbox into the AddPrevClassStudents component
